Backend GetAllServerData handler function fixed
Before the GetAllServerData only returned servers which had server data registered.

In the case where an RCON connection can not be established to the server, this would mean that the server would not be visible to client as it would not be returned by this function.

This was resolved by looping through all the servers before checking if they have server data registered, and if they don't an empty server data field is created for them with Offline set to false.